device/requirements.txt
iotc==1.1.1
python-dotenv==0.19.2
requests==2.26.0
pyserial